+++ test/bugwithbaad.R
@@ -1,4 +1,0 @@
-
-dat <- read.csv("test/baaddf.csv")
-
-sm <- sma(y ~ x*Group, data=dat, quiet=TRUE, log="xy")